backend/pyproject.toml
[project]
name = "sql-agent-backend"
version = "0.1.0"
description = "Distribution analytics AI agent backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "langgraph>=0.0.26",
    "sqlalchemy>=2.0.28",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "pydantic>=2.6.4",
    "typing-extensions>=4.10.0",
    "openai>=1.12.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
